mlops/02-sm_pipeline_basic_example/requirements.txt
scikit-learn
xgboost==2.1.4
s3fs==0.4.2
sagemaker
pandas
gevent
geventhttpclient
mlflow-skinny==2.16.2
sagemaker-mlflow
